fix: sym & delta argv

+++ examples/har_classifier/train_c_delta.c
@@ -517,18 +517,18 @@ int main(int argc, char *argv[]) {
         return 1;
     }
 
-    int trial_number = atof(argv[1]);
+    int trial_number = atoi(argv[1]);
     int batch = 64;
 
-    size_t delta_reduction = 0;
+    int delta_reduction = 0;
     if (argc > 2) {
-        trial_number = atof(argv[1]);
+        trial_number = atoi(argv[1]);
         delta_reduction = atoi(argv[2]);
         g_deltaBits = g_symBits - delta_reduction;
         g_lr = atof(argv[3]);
         g_momentum = atof(argv[4]);
-        g_epochs = atof(argv[5]);
-        batch = atof(argv[6]);
+        g_epochs = atoi(argv[5]);
+        batch = atoi(argv[6]);
         //rounding_mode = atof(argv[7]);
     }
     int len = snprintf(NULL, 0, "examples/har_classifier/logs/with_deltas/delta_reduction_%dtrial_%d.json", delta_reduction, trial_number);
